+++ CHANGELOG.md
@@ -3,6 +3,25 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.7.0] - 15/08/2025
+
+### `Added`
+
+- Singularity image build script for OSX systems using Docker.
+- Singularity definition file for building the AE Singularity image.
+- `scripts/cron_daily_prepare.sh`: 
+  - Now uses the new Singularity image `sidora_AE_singularity.sif` for running the `prepare_eager_tsv.R` script.
+- `scripts/update_poseidon_packages.sh`: 
+  - Now uses the new Singularity image `sidora_AE_singularity.sif` for running the `fill_in_janno.R` and `update_dataset_from_janno.R` scripts.
+
+### `Fixed`
+
+### `Dependencies`
+
+- Now requires the singularity image `sidora_AE_singularity.sif`, which can be built from the provided definition file and build script.
+
+### `Deprecated`
+
 ## [1.6.2] - 29/07/2025
 
 ### `Added`